pyproject.toml
[tool.poetry]
name = "streamlit-sb-docs"
version = "0.1.0"
description = ""
authors = ["mharary <moshe.harary@skyboxsecurity.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.9.7 || >3.9.7,<4.0"
openai = "^1.53.0"
langchain-openai = "^0.2.4"
certifi = "^2024.8.30"
streamlit = "^1.39.0"
st-files-connection = "^0.1.0"
gcsfs = "^2024.10.0"
google-cloud-storage = "^2.18.2"
pypdf2 = "^3.0.1"
tiktoken = "^0.8.0"
pinecone = "^5.3.1"
jinja2 = "^3.1.4"
nltk = "^3.9.1"
rank-bm25 = "^0.2.2"
plotly = "^5.24.1"
streamlit-date-picker = "^0.0.3"
sqlitecloud = "^0.0.83"
langchain-core = "^0.3.18"
langchain = "^0.3.7"
langchain-community = "^0.3.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
